requirements.txt
flask~=2.2.5
google-generativeai
gunicorn